Remove WriteCompare in PWM constructor
The PWM constructors in the PSoC Creator project default the Compare value to 4500. Since the Python API PWM constructor sets this explicitly, there is no danger in removing it from the Python API. In addition, it is important to explicitly not set it in the constructor so that multiple instances of a PWM object do not overwrite the running behavior.